requirements.txt
wheel
Flask
requests
prometheus_client
uwsgi
pyyaml
